ali up

+++ frontend/src/module/user-center/ali-cloud/component/ali-sub-list.component.ts
@@ -24,7 +24,9 @@ export class AliSubListComponent implements OnInit{
         private router : Router,
         private service : AliSubService,
         private layoutService : LayoutService,
-        private activatedRouter : ActivatedRoute
+        private activatedRouter : ActivatedRoute,
+        private v:Validation,
+        private validationService: ValidationService
     ) {
         if (activatedRouter.snapshot.params["loginName"]) {
             this.name = activatedRouter.snapshot.params["loginName"] || "";
@@ -50,24 +52,25 @@ export class AliSubListComponent implements OnInit{
     name: string;
     type: string;
     data: Array<AliSubList>;
-    id: string;
+    Majorid: string;
     subInfo: AliSubList= new AliSubList();
     departsList: Array<DepartList>;
     selectedDepartment: string;
     selectedDepartmentId: string;
     testInfo: string;
+    id: string;
 
     ngOnInit (){
         console.log('init');
         this.activatedRouter.params.forEach((params: Params) =>{
-            this.id= params["id"];
+            this.Majorid= params["id"];
         })
         this.getData();
     }
 
     getData() {
         this.layoutService.show();
-        this.service.getData(this.id)
+        this.service.getData(this.Majorid)
             .then(
                 response => {
                     this.layoutService.hide();
@@ -124,12 +127,15 @@ export class AliSubListComponent implements OnInit{
 
     crePage(){
         this.type= "create";
+        this.testInfo= "";
         this.getDepartsList();
         this.subMng.open("创建子账号");
     }
 
     editPage(item){
         this.type= "edit";
+        this.id= item.id;
+        this.testInfo= "";
         this.layoutService.show();
         this.service.getDetail(item.id)
             .then(
@@ -165,8 +171,23 @@ export class AliSubListComponent implements OnInit{
 
     distriPage(item){
         this.type= "distribute";
-        this.getDepartsList();
-        this.distriDepart.open("分配部门")
+        this.selectedDepartment= item.departmentName;
+        this.selectedDepartmentId= "";
+        this.id= item.id;
+        this.layoutService.show();
+        this.service.getDepartsList()
+            .then(
+                response => {
+                    this.layoutService.hide();
+                    if (response && 100 == response["resultCode"]) {
+                        this.departsList = response["resultContent"];
+                        this.distriDepart.open("分配部门")
+                    } else {
+                        this.showAlert("COMMON.OPERATION_ERROR");
+                    }
+                }
+            )
+            .catch((e) => this.onRejected(e));
     }
 
     selected(item: DepartList){
@@ -175,6 +196,7 @@ export class AliSubListComponent implements OnInit{
         });
         item.selected= true;
         this.selectedDepartment= item.departmentName;
+        this.selectedDepartmentId= item.id;
     }
 
     reset(){
@@ -182,10 +204,11 @@ export class AliSubListComponent implements OnInit{
             p.selected= false;
         });
         this.selectedDepartment= "";
+        this.selectedDepartmentId= "";
     }
 
     enable(item){
-        this.confirm.open("启用账号","您选择启用，请确认");
+        this.confirm.open("USER_CENTER.ENABLE_ALIACCOUNT","USER_CENTER.YOU_SELECT_ENABLE_ALIACCOUNT^^^"+item.loginName);
         this.confirm.ccf= ()=>{
             this.layoutService.show();
             this.service.enable(item.id)
@@ -205,7 +228,7 @@ export class AliSubListComponent implements OnInit{
     }
 
     disable(item){
-        this.confirm.open("禁用账号","您选择禁用，请确认");
+        this.confirm.open("USER_CENTER.DISABKE_ALIACCOUNT","USER_CENTER.YOU_SELECT_DISABLE_ALIACCOUNT^^^"+item.loginName);
         this.confirm.ccf= ()=>{
             this.layoutService.show();
             this.service.disable(item.id)
@@ -224,30 +247,39 @@ export class AliSubListComponent implements OnInit{
     }
 
     delete(item){
-        this.confirm.open("删除账号","您选择删除，请确认");
-        this.confirm.ccf= ()=>{
-            this.layoutService.show();
-            this.service.delete(item.id)
-                .then(
-                    response => {
-                        this.layoutService.hide();
-                        if (response && 100 == response["resultCode"]) {
-                            this.getData();
-                        } else {
-                            this.showAlert("COMMON.OPERATION_ERROR");
+        if(item.status == 2 && item.departId== ""){
+            this.confirm.open("USER_CENTER.DELETE_ALIACCOUNT","USER_CENTER.YOU_SELECT_DELETE_ALIACCOUNT^^^"+item.loginName);
+            this.confirm.ccf= ()=>{
+                this.layoutService.show();
+                this.service.delete(item.id)
+                    .then(
+                        response => {
+                            this.layoutService.hide();
+                            if (response && 100 == response["resultCode"]) {
+                                this.getData();
+                            } else {
+                                this.showAlert("COMMON.OPERATION_ERROR");
+                            }
                         }
-                    }
-                )
-                .catch((e) => this.onRejected(e));
+                    )
+                    .catch((e) => this.onRejected(e));
+            }
+        }else{
+            this.showAlert("必须为禁用状态且未分配给任何部门时才能删除");
         }
+
     }
 
     operate(){
         if(this.type== "info"){
             this.subMng.close();
         }else if(this.type== "create"){
+            let selectedDepart= this.departsList.find((p)=>{
+                return p.id == this.subInfo.departId;
+            });
+            this.subInfo.departmentName= selectedDepart && selectedDepart.departmentName || "";
             this.layoutService.show();
-            this.service.create(this.id, this.subInfo)
+            this.service.create(this.Majorid, this.subInfo)
                 .then(
                     response => {
                         this.layoutService.hide();
@@ -261,10 +293,48 @@ export class AliSubListComponent implements OnInit{
                 )
                 .catch((e) => this.onRejected(e));
         }else if(this.type== "edit"){
-
+            this.layoutService.show();
+            this.service.edit(this.id, this.subInfo)
+                .then(
+                    response => {
+                        this.layoutService.hide();
+                        if (response && 100 == response["resultCode"]) {
+                            this.getData();
+                            this.subMng.close();
+                            console.log("edit后",this.subInfo);
+                        }else {
+                            this.showAlert("COMMON.OPERATION_ERROR");
+                        }
+                    }
+                )
+                .catch((e) => this.onRejected(e));
         }else{
-
+            this.layoutService.show();
+            this.service.editDepart(this.id, this.selectedDepartmentId)
+                .then(
+                    response => {
+                        this.layoutService.hide();
+                        if (response && 100 == response["resultCode"]) {
+                            this.getData();
+                            this.distriDepart.close();
+                            console.log("editDepart",this.id, this.selectedDepartmentId);
+                        }else {
+                            this.showAlert("COMMON.OPERATION_ERROR");
+                        }
+                    }
+                )
+                .catch((e) => this.onRejected(e));
         }
+    }
+
+    checkForm(key?:string){
+        const regs:ValidationRegs = {
+            accessKey: [this.subInfo.accessKey, [this.v.isUnBlank], "accessKey不能为空"],
+            accessSecret: [this.subInfo.accessSecret, [this.v.isUnBlank], "accessSecret不能为空"],
+            loginName: [this.subInfo.loginName, [this.v.isBase, this.v.isUnBlank], "loginName不能为空"],
+        }
+
+        return this.v.check(key, regs);
     }
 
     showAlert(msg: string): void {
